POC to get condition-variable.obj building without windows.h
This is a test to see what is involved in getting v8 to build without
Windows header files being included from win32-headers.h. Some issues
have been ignored for now (CRITICAL_SECTION) but no issues that have not
previously been solved in Chromium have been found. With this change I
can build obj/v8/v8_libbase/condition-variable.obj without including
windows.h from win32-headers.h.

Bug: 796644

+++ src/base/bits.h
@@ -194,7 +194,7 @@ inline size_t RoundUpToPowerOfTwo(size_t value) {
   if (sizeof(size_t) == sizeof(uint64_t)) {
     return RoundUpToPowerOfTwo64(value);
   } else {
-    return RoundUpToPowerOfTwo32(value);
+    return RoundUpToPowerOfTwo32((unsigned int)value);
   }
 }
